+++ src/main/java/org/aero/mtip/util/MtipUtils.java
@@ -283,6 +283,8 @@ public class MtipUtils {
       return UmlConstants.ARTIFACT;
     } else if (SysML.isBoundReference(element)) {
       return SysmlConstants.BOUND_REFERENCE;
+    } else if (SysML.isBusinessRequirement(element)) {
+      return SysmlConstants.BUSINESS_REQUIREMENT;
     } else if (SysML.isClassifierBehavior(element)) {
       return SysmlConstants.CLASSIFIER_BEHAVIOR_PROPERTY;
     } else if (element instanceof CallBehaviorAction) {
@@ -660,6 +662,16 @@ public class MtipUtils {
     }
 
     return SysmlConstants.PACKAGE;
+  }
+  
+  public static boolean isReferencedElement(Element element) {
+    Project project = Project.getProject(element);
+    
+    if (Application.getInstance().getProject().equals(project)) {
+      return false;
+    }
+    
+    return true;
   }
 
   public static boolean isRelationship(Element element) {